fixed authorization

+++ api/controllers/categoryController.js
@@ -1,4 +1,6 @@
 const Category = require('../../models/category');
+const jwt = require('jsonwebtoken');
+// const authAdmin = require('../middleware/adminAuth');
 
 module.exports = {
     root: function(req, res, next) {
@@ -8,12 +10,37 @@ module.exports = {
             .catch(err => console.log(err));
     },
     new: function(req, res, next) {
-        res.render('category/new');
+        jwt.verify(req.token, "thesecret", (err, user) => {
+            if (err) {
+                return res.sendStatus(403);
+            } else {
+                req.user = user;
+            }
+        });
+        if (req.user.name.role != 'admin') {
+            return res.sendStatus(403);
+        } else {
+            res.render('category/new');
+        }
+        
     },
     createCategory: function(req, res, next) {
-        Category.create(req.body.category)
+        jwt.verify(req.token, "thesecret", (err, user) => {
+            if (err) {
+                return res.sendStatus(403);
+            } else {
+                // res.render('category/new');
+                req.user = user;
+            }
+        });
+
+        if (req.user.name.role != 'admin') {
+            return res.sendStatus(403);
+        } else {
+            Category.create(req.body.category)
             .then(category => res.send('successfully added category'))
             .catch(err => console.log(err));
+        }
     },
     show: function(req, res, next) {
         Category.findById(req.params.id)
@@ -21,18 +48,60 @@ module.exports = {
             .catch(err => console.log(err));
     },
     edit: function(req, res, next) {
-        Category.findById(req.params.id)
+        jwt.verify(req.token, "thesecret", (err, user) => {
+            if (err) {
+                return res.sendStatus(403);
+            } else {
+                // res.render('category/new');
+                req.user = user;
+            }
+        });
+
+        if (req.user.name.role != 'admin') {
+            return res.sendStatus(403);
+        } else {
+            Category.findById(req.params.id)
             .then(category => res.render('category/edit', {category: category}))
             .catch(err => console.log(err));
+        }
+        
     },
     update: function(req, res, next) {
-        Category.findByIdAndUpdate(req.params.id, req.body.category)
-            .then(() => res.redirect('/categories/' + req.params.id))
-            .catch(err => console.log(err));
+        jwt.verify(req.token, "thesecret", (err, user) => {
+            if (err) {
+                return res.sendStatus(403);
+            } else {
+                res.render('category/new');
+                req.user = user;
+            }
+        });
+
+        if (req.user.name.role != 'admin') {
+            return res.sendStatus(403);
+        } else {
+            Category.findByIdAndUpdate(req.params.id, req.body.category)
+                .then(() => res.redirect('/categories/' + req.params.id))
+                .catch(err => console.log(err));
+        }
+        
     },
     delete: function(req, res, next) {
-        Category.findByIdAndDelete(req.params.id)
+        jwt.verify(req.token, "thesecret", (err, user) => {
+            if (err) {
+                return res.sendStatus(403);
+            } else {
+                res.render('category/new');
+                req.user = user;
+            }
+        });
+
+        if (req.user.name.role != 'admin') {
+            return res.sendStatus(403);
+        } else {
+            Category.findByIdAndDelete(req.params.id)
             .then(() => res.redirect('/categories'))
             .catch(err => console.log(err));
+        }
+        
     }
 }
